requirements.txt
awsebcli==3.12.1
blessed==1.14.2
botocore==1.8.26
cement==2.8.2
certifi==2017.11.5
chardet==3.0.4
colorama==0.3.7
decorator==4.1.2
diff-match-patch==20121119
Django==1.11.6
django-cors-headers==2.2.0
django-filter==1.0.4
django-import-export==1.0.0
django-material==1.1.1
django-viewflow==1.1.0
docker-py==1.7.2
dockerpty==0.4.1
docopt==0.6.2
docutils==0.14
enum34==1.1.6
et-xmlfile==1.0.1
functools32==3.2.3.post2
gunicorn==19.7.1
idna==2.6
ipython-genutils==0.2.0
jdcal==1.3
jmespath==0.9.3
jsonschema==2.6.0
jupyter-core==4.4.0
nbformat==4.4.0
odfpy==1.3.6
openpyxl==2.5.1
pathspec==0.5.0
plotly==2.2.3
PyJWT==1.5.3
python-dateutil==2.6.1
pytz==2017.2
PyYAML==3.12
requests==2.9.1
semantic-version==2.5.0
six==1.11.0
tablib==0.12.1
tabulate==0.7.5
termcolor==1.1.0
traitlets==4.3.2
unicodecsv==0.14.1
urllib3==1.22
wcwidth==0.1.7
websocket-client==0.46.0
whitenoise==3.3.1
xlrd==1.1.0
xlwt==1.3.0
